# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go Web API Boilerplate
  version: 1.0.0
  description: OpenAPI specification for the v1 HTTP API.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /inventory/{id}:
    put:
      tags: [Inventory]
      summary: Set inventory stock for product
      operationId: upsertInventoryForProduct
      parameters:
        - in: path
          name: id
          required: true
          description: Product UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertInventoryRequest'
      responses:
        '200':
          description: Inventory updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpsertInventoryResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /cart/items:
    post:
      tags: [Cart]
      summary: Add item to active cart
      operationId: createCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCartItemRequest'
      responses:
        '200':
          description: Cart item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCartItemResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /orders:
    post:
      tags: [Orders]
      summary: Create order from active cart
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '402':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      operationId: updateOrderStatus
      parameters:
        - in: path
          name: id
          required: true
          description: Order UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /payment-methods:
    post:
      tags: [Payments]
      summary: Create payment method
      operationId: createPaymentMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentMethodRequest'
      responses:
        '201':
          description: Payment method created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
    get:
      tags: [Payments]
      summary: List payment methods
      operationId: listPaymentMethods
      responses:
        '200':
          description: Payment methods listed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethodResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /orders/{id}/pay:
    post:
      tags: [Payments]
      summary: Pay order
      operationId: payOrder
      parameters:
        - in: path
          name: id
          required: true
          description: Order UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayOrderRequest'
      responses:
        '200':
          description: Order paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayOrderResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '402':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
  responses:
    ErrorResponse:
      description: API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UpsertInventoryRequest:
      type: object
      required: [in_stock]
      properties:
        in_stock:
          type: integer
          format: int32
          minimum: 0
    UpsertInventoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        in_stock:
          type: integer
          format: int32
        reserved:
          type: integer
          format: int32
    CreateCartItemRequest:
      type: object
      required: [user_uid, product_id, quantity]
      properties:
        user_uid:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          minimum: 1
        note:
          type: string
        prescription_ref:
          type: string
    CreateCartItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cart_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
        unit_price:
          type: integer
          format: int64
    CreateOrderLocation:
      type: object
      properties:
        address_line:
          type: string
        city:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
    CreateOrderRequest:
      type: object
      required: [cart_id, payment_method_id, fulfillment_type]
      properties:
        cart_id:
          type: string
          format: uuid
        payment_method_id:
          type: string
          description: Use "cash" or a payment method UUID.
        fulfillment_type:
          type: string
          enum: [pickup, delivery]
        location:
          $ref: '#/components/schemas/CreateOrderLocation'
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        cart_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, confirmed, cancelled, out_for_delivery, completed, refunded]
        fulfillment_type:
          type: string
          enum: [pickup, delivery]
        subtotal:
          type: integer
          format: int64
        tax:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, out_for_delivery, completed, refunded]
    CreatePaymentMethodRequest:
      type: object
      required: [type, label]
      properties:
        type:
          type: string
          enum: [cash, card]
        label:
          type: string
        is_default:
          type: boolean
    PaymentMethodResponse:
      type: object
      properties:
        id:
          type: string
          description: Returns "cash" for default cash method, otherwise UUID string.
        type:
          type: string
          enum: [cash, card]
        label:
          type: string
        is_default:
          type: boolean
    PayOrderRequest:
      type: object
      required: [payment_method_id, amount]
      properties:
        payment_method_id:
          type: string
          description: Use "cash" or a payment method UUID.
        amount:
          type: integer
          format: int64
          minimum: 1
    PayOrderResponse:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, confirmed, cancelled, out_for_delivery, completed, refunded]
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: auth/unauthorized
        errors:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
            - type: object
              additionalProperties: true