# generated/typescript-fetch/models/CreateCartItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateCartItemResponse
 */
export interface CreateCartItemResponse {
    /**
     * 
     * @type {string}
     * @memberof CreateCartItemResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateCartItemResponse
     */
    cartId?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateCartItemResponse
     */
    productId?: string;
    /**
     * 
     * @type {number}
     * @memberof CreateCartItemResponse
     */
    quantity?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateCartItemResponse
     */
    unitPrice?: number;
}

/**
 * Check if a given object implements the CreateCartItemResponse interface.
 */
export function instanceOfCreateCartItemResponse(value: object): value is CreateCartItemResponse {
    return true;
}

export function CreateCartItemResponseFromJSON(json: any): CreateCartItemResponse {
    return CreateCartItemResponseFromJSONTyped(json, false);
}

export function CreateCartItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateCartItemResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'cartId': json['cart_id'] == null ? undefined : json['cart_id'],
        'productId': json['product_id'] == null ? undefined : json['product_id'],
        'quantity': json['quantity'] == null ? undefined : json['quantity'],
        'unitPrice': json['unit_price'] == null ? undefined : json['unit_price'],
    };
}

export function CreateCartItemResponseToJSON(json: any): CreateCartItemResponse {
    return CreateCartItemResponseToJSONTyped(json, false);
}

export function CreateCartItemResponseToJSONTyped(value?: CreateCartItemResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'cart_id': value['cartId'],
        'product_id': value['productId'],
        'quantity': value['quantity'],
        'unit_price': value['unitPrice'],
    };
}