# generated/typescript-fetch/models/CreateOrderLocation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateOrderLocation
 */
export interface CreateOrderLocation {
    /**
     * 
     * @type {string}
     * @memberof CreateOrderLocation
     */
    addressLine?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateOrderLocation
     */
    city?: string;
    /**
     * 
     * @type {number}
     * @memberof CreateOrderLocation
     */
    lat?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateOrderLocation
     */
    lng?: number;
}

/**
 * Check if a given object implements the CreateOrderLocation interface.
 */
export function instanceOfCreateOrderLocation(value: object): value is CreateOrderLocation {
    return true;
}

export function CreateOrderLocationFromJSON(json: any): CreateOrderLocation {
    return CreateOrderLocationFromJSONTyped(json, false);
}

export function CreateOrderLocationFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateOrderLocation {
    if (json == null) {
        return json;
    }
    return {
        
        'addressLine': json['address_line'] == null ? undefined : json['address_line'],
        'city': json['city'] == null ? undefined : json['city'],
        'lat': json['lat'] == null ? undefined : json['lat'],
        'lng': json['lng'] == null ? undefined : json['lng'],
    };
}

export function CreateOrderLocationToJSON(json: any): CreateOrderLocation {
    return CreateOrderLocationToJSONTyped(json, false);
}

export function CreateOrderLocationToJSONTyped(value?: CreateOrderLocation | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'address_line': value['addressLine'],
        'city': value['city'],
        'lat': value['lat'],
        'lng': value['lng'],
    };
}