# generated/typescript-fetch/models/CreatePaymentMethodRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePaymentMethodRequest
 */
export interface CreatePaymentMethodRequest {
    /**
     * 
     * @type {CreatePaymentMethodRequestTypeEnum}
     * @memberof CreatePaymentMethodRequest
     */
    type: CreatePaymentMethodRequestTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CreatePaymentMethodRequest
     */
    label: string;
    /**
     * 
     * @type {boolean}
     * @memberof CreatePaymentMethodRequest
     */
    isDefault?: boolean;
}


/**
 * @export
 */
export const CreatePaymentMethodRequestTypeEnum = {
    Cash: 'cash',
    Card: 'card'
} as const;
export type CreatePaymentMethodRequestTypeEnum = typeof CreatePaymentMethodRequestTypeEnum[keyof typeof CreatePaymentMethodRequestTypeEnum];


/**
 * Check if a given object implements the CreatePaymentMethodRequest interface.
 */
export function instanceOfCreatePaymentMethodRequest(value: object): value is CreatePaymentMethodRequest {
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('label' in value) || value['label'] === undefined) return false;
    return true;
}

export function CreatePaymentMethodRequestFromJSON(json: any): CreatePaymentMethodRequest {
    return CreatePaymentMethodRequestFromJSONTyped(json, false);
}

export function CreatePaymentMethodRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePaymentMethodRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'type': json['type'],
        'label': json['label'],
        'isDefault': json['is_default'] == null ? undefined : json['is_default'],
    };
}

export function CreatePaymentMethodRequestToJSON(json: any): CreatePaymentMethodRequest {
    return CreatePaymentMethodRequestToJSONTyped(json, false);
}

export function CreatePaymentMethodRequestToJSONTyped(value?: CreatePaymentMethodRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'type': value['type'],
        'label': value['label'],
        'is_default': value['isDefault'],
    };
}