# generated/typescript-fetch/models/PayOrderRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PayOrderRequest
 */
export interface PayOrderRequest {
    /**
     * Use "cash" or a payment method UUID.
     * @type {string}
     * @memberof PayOrderRequest
     */
    paymentMethodId: string;
    /**
     * 
     * @type {number}
     * @memberof PayOrderRequest
     */
    amount: number;
}

/**
 * Check if a given object implements the PayOrderRequest interface.
 */
export function instanceOfPayOrderRequest(value: object): value is PayOrderRequest {
    if (!('paymentMethodId' in value) || value['paymentMethodId'] === undefined) return false;
    if (!('amount' in value) || value['amount'] === undefined) return false;
    return true;
}

export function PayOrderRequestFromJSON(json: any): PayOrderRequest {
    return PayOrderRequestFromJSONTyped(json, false);
}

export function PayOrderRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PayOrderRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'paymentMethodId': json['payment_method_id'],
        'amount': json['amount'],
    };
}

export function PayOrderRequestToJSON(json: any): PayOrderRequest {
    return PayOrderRequestToJSONTyped(json, false);
}

export function PayOrderRequestToJSONTyped(value?: PayOrderRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'payment_method_id': value['paymentMethodId'],
        'amount': value['amount'],
    };
}