# generated/typescript-fetch/models/PayOrderResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PayOrderResponse
 */
export interface PayOrderResponse {
    /**
     * 
     * @type {string}
     * @memberof PayOrderResponse
     */
    orderId?: string;
    /**
     * 
     * @type {PayOrderResponseStatusEnum}
     * @memberof PayOrderResponse
     */
    status?: PayOrderResponseStatusEnum;
}


/**
 * @export
 */
export const PayOrderResponseStatusEnum = {
    Pending: 'pending',
    Confirmed: 'confirmed',
    Cancelled: 'cancelled',
    OutForDelivery: 'out_for_delivery',
    Completed: 'completed',
    Refunded: 'refunded'
} as const;
export type PayOrderResponseStatusEnum = typeof PayOrderResponseStatusEnum[keyof typeof PayOrderResponseStatusEnum];


/**
 * Check if a given object implements the PayOrderResponse interface.
 */
export function instanceOfPayOrderResponse(value: object): value is PayOrderResponse {
    return true;
}

export function PayOrderResponseFromJSON(json: any): PayOrderResponse {
    return PayOrderResponseFromJSONTyped(json, false);
}

export function PayOrderResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PayOrderResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'orderId': json['order_id'] == null ? undefined : json['order_id'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function PayOrderResponseToJSON(json: any): PayOrderResponse {
    return PayOrderResponseToJSONTyped(json, false);
}

export function PayOrderResponseToJSONTyped(value?: PayOrderResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'order_id': value['orderId'],
        'status': value['status'],
    };
}